+++ pages/Dashboard/api/environment.ts
@@ -1,6 +1,6 @@
 import type { AxiosResponse } from "axios";
 import { api } from "@/lib/axios";
-import type { IParamsEnvironment } from "@/interfaces/common";
+import type { IParamsEnvironment } from "@/pages/Dashboard/interfaces/environment.ts";
 
 export default {
   readEnvironment: (
