+++ pages/Dashboard/components/ThePiaChartAbsolute.vue
@@ -46,7 +46,7 @@
 <script setup lang="ts">
 import type { FormRules } from "element-plus";
 import type { IChartPie } from "@/interfaces/chart";
-import type { IParamsEnvironment } from "@/interfaces/common";
+import type { IParamsEnvironment } from "@/pages/Dashboard/interfaces/environment.ts";
 import { countries, elements } from "@/constants";
 import { useDashboard } from "@/pages/Dashboard/composables/dashboard";
 import {
